cci_datatypes: split into cci_core_types.h and cci_mutable_types.h
