+++ src/cci_cfg/cci_param_untyped_handle.h
@@ -22,7 +22,7 @@
 
 #include <string>
 #include "cci_cfg/cci_config_macros.h"
-#include "cci_cfg/cci_datatypes.h"
+#include "cci_cfg/cci_core_types.h"
 #include "cci_cfg/cci_param_if.h"
 #include "cci_cfg/cci_value.h"
