+++ src/cci_cfg/cci_value.h
@@ -20,7 +20,7 @@
 #ifndef CCI_CCI_VALUE_H_INCLUDED_
 #define CCI_CCI_VALUE_H_INCLUDED_
 
-#include "cci_cfg/cci_datatypes.h"
+#include "cci_cfg/cci_core_types.h"
 #include "cci_core/systemc.h" // sc_dt::(u)int64, potentially strip out
 
 #include <cstring> // std::strlen
