+++ src/cci_configuration
@@ -28,7 +28,8 @@
 // CCI config includes
 
 #include "cci_cfg/cci_debug.h"
-#include "cci_cfg/cci_datatypes.h"
+#include "cci_cfg/cci_core_types.h"
+#include "cci_cfg/cci_mutable_types.h"
 #include "cci_cfg/cci_originator.h"
 #include "cci_cfg/cci_broker_manager.h"
 #include "cci_cfg/cci_value.h"
